pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sis_etl_tool"
version = "1.0.0"
authors = [{name = "Shan Peiris", email = "shan.peiris@myblueprint.ca"}]
description = "SIS Data ETL Tool for myBlueprint - SpacesEDU"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "pandas>=1.4.0",
    "PyYAML>=6.0",
    "python-dateutil>=2.8.2",
    "pyinstaller>=5.0",
]
[tool.poetry.scripts]
sis-etl = "src.main:main"